# server/docs/swagger/swagger.yaml
basePath: /
definitions:
...
  github_com_ix-pay_ixpay-pro_internal_dto_base_request.BatchDeleteOperationLogRequest:
    properties:
      ids:
        items:
          type: string
        type: array
    required:
    - ids
    type: object
...
  github_com_ix-pay_ixpay-pro_internal_dto_base_request.ClearOperationLogByTimeRangeRequest:
    properties:
      endTime:
        type: string
      startTime:
        type: string
    required:
    - endTime
    - startTime
    type: object
...
  github_com_ix-pay_ixpay-pro_internal_dto_base_response.OperationLogListResponse:
    properties:
      list:
        items:
          $ref: '#/definitions/github_com_ix-pay_ixpay-pro_internal_dto_base_response.OperationLogResponse'
        type: array
      page:
        type: integer
      pageSize:
        type: integer
      total:
        type: integer
    type: object
  github_com_ix-pay_ixpay-pro_internal_dto_base_response.OperationLogResponse:
    properties:
      clientIp:
        type: string
      createdAt:
        type: string
      description:
        type: string
      duration:
        type: integer
      errorMessage:
        type: string
      id:
        example: ""
        type: string
      isSuccess:
        type: boolean
      method:
        type: string
      module:
        type: string
      nickname:
        type: string
      operationType:
        type: integer
      params:
        type: string
      path:
        type: string
      result:
        type: string
      statusCode:
        type: integer
      updatedAt:
        type: string
      userAgent:
        type: string
      userId:
        example: ""
        type: string
      username:
        type: string
    type: object
  github_com_ix-pay_ixpay-pro_internal_dto_base_response.OperationLogStatisticsResponse:
    properties:
      avgDuration:
        type: integer
      failedCount:
        type: integer
      operationTypes:
        items:
          properties:
            count:
              type: integer
            type:
              type: integer
          type: object
        type: array
      successCount:
        type: integer
      successRate:
        type: integer
      totalCount:
        type: integer
    type: object
...
  github_com_ix-pay_ixpay-pro_internal_utils_common_baseRes.Response:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
host: 127.0.0.1:8586
info:
  contact:
    email: support@example.com
    name: API Support
  description: 基于Gin和PostgreSQL的微信支付API服务
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: 微信支付API服务
  version: "1.0"
paths:
...
  /api/admin/logs:
    get:
      consumes:
      - application/json
      description: 获取系统操作日志列表，支持分页和多条件过滤
      parameters:
      - description: 页码
        in: query
        name: page
        required: true
        type: integer
      - description: 每页数量
        in: query
        name: pageSize
        required: true
        type: integer
      - description: 开始时间 (格式:2006-01-02)
        in: query
        name: startTime
        type: string
      - description: 结束时间 (格式:2006-01-02)
        in: query
        name: endTime
        type: string
      - description: 用户名
        in: query
        name: username
        type: string
      - description: 操作模块
        in: query
        name: module
        type: string
      - description: 操作类型
        in: query
        name: operationType
        type: integer
      - description: 操作结果
        in: query
        name: isSuccess
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: 操作日志列表
          schema:
            allOf:
            - $ref: '#/definitions/github_com_ix-pay_ixpay-pro_internal_utils_common_baseRes.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_ix-pay_ixpay-pro_internal_dto_base_response.OperationLogListResponse'
              type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 获取操作日志列表
      tags:
      - 系统管理
  /api/admin/logs/{id}:
    delete:
      consumes:
      - application/json
      description: 根据日志ID删除指定的操作日志
      parameters:
      - description: 日志ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 根据ID删除操作日志
      tags:
      - 系统管理
    get:
      consumes:
      - application/json
      description: 根据日志 ID 获取详细的操作日志信息
      parameters:
      - description: 日志 ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 操作日志详情
          schema:
            allOf:
            - $ref: '#/definitions/github_com_ix-pay_ixpay-pro_internal_utils_common_baseRes.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_ix-pay_ixpay-pro_internal_dto_base_response.OperationLogResponse'
              type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: 日志不存在
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 根据 ID 获取操作日志
      tags:
      - 系统管理
  /api/admin/logs/batch-delete:
    post:
      consumes:
      - application/json
      description: 批量删除指定 ID 列表的操作日志
      parameters:
      - description: 批量删除请求参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_ix-pay_ixpay-pro_internal_dto_base_request.BatchDeleteOperationLogRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 批量删除成功
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 批量删除操作日志
      tags:
      - 系统管理
  /api/admin/logs/clear:
    post:
      consumes:
      - application/json
      description: 清空指定时间范围的操作日志
      parameters:
      - description: 清空日志请求参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_ix-pay_ixpay-pro_internal_dto_base_request.ClearOperationLogByTimeRangeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 清空成功
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 根据时间范围清空操作日志
      tags:
      - 系统管理
  /api/admin/logs/statistics:
    get:
      consumes:
      - application/json
      description: 获取操作日志的统计信息，包括操作类型分布等
      parameters:
      - description: 开始时间 (格式:2006-01-02)
        in: query
        name: startTime
        type: string
      - description: 结束时间 (格式:2006-01-02)
        in: query
        name: endTime
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 操作日志统计信息
          schema:
            allOf:
            - $ref: '#/definitions/github_com_ix-pay_ixpay-pro_internal_utils_common_baseRes.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_ix-pay_ixpay-pro_internal_dto_base_response.OperationLogStatisticsResponse'
              type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 获取操作日志统计信息
      tags:
      - 系统管理
...
securityDefinitions:
  BearerAuth:
    in: header [Bearer ]
    name: Authorization
    type: apiKey
swagger: "2.0"